[PREINVEST][REWARDS] fix bug silder

+++ mepa-front/src/main/java/fr/epita/sigl/mepa/front/controller/InsertDummyController.java
@@ -14,11 +14,9 @@ import fr.epita.sigl.mepa.core.service.RewardService;
 import fr.epita.sigl.mepa.front.controller.home.HomeController;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import org.springframework.asm.Label;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.ModelMap;
-import org.springframework.web.bind.annotation.ModelAttribute;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.SessionAttributes;
 
@@ -107,7 +105,7 @@ public class InsertDummyController {
             Reward r = new Reward();
             r.setName("Coucou");
             r.setDescription("This is a description");
-            r.setCostStart((long) 10);
+            r.setCostStart((float) 10);
             rewardService.createReward(r);
         }
 
@@ -138,7 +136,7 @@ public class InsertDummyController {
                 Reward r = new Reward();
                 r.setName("Coucou");
                 r.setDescription("This is a description");
-                r.setCostStart((long) 10);
+                r.setCostStart((float) 10);
                 this.rewardService.createReward(r);
                 rewards.add(r);
             }
@@ -174,7 +172,7 @@ public class InsertDummyController {
                 Reward r = new Reward();
                 r.setName("Gildas est beau" + j);
                 r.setDescription("This is a description");
-                r.setCostStart((long) 10);
+                r.setCostStart((float) 10);
                 this.rewardService.createReward(r);
                 rewards.add(r);
             }
